Fix Allegro crashes/compilation under 64 bit Windows.

+++ src/win/wsystem.c
@@ -491,7 +491,7 @@ int _WinMain(void *_main, void *hInst, void *hPrev, char *Cmd, int nShow)
 
    /* can't use parameter because it doesn't include the executable name */
    cmdline = GetCommandLine();
-   i = strlen(cmdline) + 1;
+   i = (int)strlen(cmdline) + 1;
    argbuf = _AL_MALLOC(i);
    memcpy(argbuf, cmdline, i);
